test: controller 테스트 클래스 및 패키지 생성
