--- src/test/java/roomescape/api/WelcomePageApiTest.java
@@ -1,22 +1,0 @@
-package roomescape.api;
-
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class WelcomePageApiTest {
-
-    @DisplayName("/로 요청이 들어오면 웰컴 페이지를 응답한다.")
-    @Test
-    void welcome() {
-        RestAssured.given().log().all()
-                .when().get("/")
-                .then().log().all()
-                .statusCode(200);
-    }
-}
