+++ src/test/java/roomescape/controller/AdminPageControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.api;
+package roomescape.controller;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.DisplayName;
@@ -9,7 +9,7 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class AdminPageApiTest {
+public class AdminPageControllerTest {
 
     @DisplayName("일단계: 홈화면 - /admin으로 요청이 들어오면 어드민 페이지를 응답한다.")
     @Test
@@ -25,6 +25,15 @@ public class AdminPageApiTest {
     void step_two() {
         RestAssured.given().log().all()
                 .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("/로 요청이 들어오면 웰컴 페이지를 응답한다.")
+    @Test
+    void welcome() {
+        RestAssured.given().log().all()
+                .when().get("/")
                 .then().log().all()
                 .statusCode(200);
     }
