+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -1,26 +1,35 @@
-package roomescape.api;
+package roomescape.controller;
 
-import static org.hamcrest.Matchers.is;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.hamcrest.core.Is.is;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
+import java.time.LocalDate;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
+import roomescape.domain.reservation.Reservation;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class ReservationApiTest {
+public class ReservationControllerTest {
 
     private static final Map<String, String> RESERVATION_BODY = Map.of(
             "name", "브라운",
             "date", "2023-08-05",
             "timeId", "1"
     );
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
 
     @BeforeEach
     void setUp() {
@@ -42,7 +51,7 @@ public class ReservationApiTest {
                 .when().get("/reservations")
                 .then().log().all()
                 .statusCode(200)
-                .body("size()", is(0));
+                .body("size()", Matchers.is(0));
     }
 
     @DisplayName("이단계: 예약 조회 - 존재하는 모든 예약과 200 OK를 응답한다.")
@@ -59,7 +68,7 @@ public class ReservationApiTest {
                 .when().get("/reservations")
                 .then().log().all()
                 .statusCode(200)
-                .body("size()", is(1));
+                .body("size()", Matchers.is(1));
     }
 
     @DisplayName("삼단계: 예약 추가 / 취소 - 예약을 추가하고, 200 OK를 응답한다.")
@@ -96,5 +105,64 @@ public class ReservationApiTest {
                 .when().delete("/reservations/3")
                 .then().log().all()
                 .statusCode(404);
+    }
+
+    @DisplayName("오단계: 데이터 조회하기")
+    @Test
+    void step_five() {
+        jdbcTemplate.update("INSERT INTO reservation (name, date, time_id) VALUES (?, ?, ?)",
+                "브라운", LocalDate.parse("2023-08-05"), 1L);
+
+        List<Reservation> reservations = RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200).extract()
+                .jsonPath().getList(".", Reservation.class);
+
+        Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+
+        assertThat(reservations.size()).isEqualTo(count);
+    }
+
+    @DisplayName("육단계: 데이터 추가 / 삭제하기")
+    @Test
+    void step_six() {
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(RESERVATION_BODY)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(200);
+
+        Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+        assertThat(count).isEqualTo(1);
+
+        RestAssured.given().log().all()
+                .when().delete("/reservations/1")
+                .then().log().all()
+                .statusCode(200);
+
+        Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(1) from reservation",
+                Integer.class);
+        assertThat(countAfterDelete).isEqualTo(0);
+    }
+
+    @DisplayName("팔단계: 예약과 시간 관리")
+    @Test
+    void step_eight() {
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(RESERVATION_BODY)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(200);
+
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(1));
     }
 }
